Split up client tests
